+++ Take Time BangPhra/Guest/RoomService.aspx.cs
@@ -15,9 +15,9 @@ namespace Take_Time_BangPhra.Guest
         private readonly string _connectionString = System.Configuration.ConfigurationManager.ConnectionStrings["TaketimeConnectionString"].ConnectionString;
         private GuestPortalService _guestPortalService;
         private code _code;
-        private int _reservationId;
+        private long _reservationId;
         private string _guestMobilePhone;
-        private int _accommodationId;
+        private byte _accommodationId;
 
         protected void Page_Load(object sender, EventArgs e)
         {
@@ -58,9 +58,9 @@ namespace Take_Time_BangPhra.Guest
             }
 
             DataRow session = dtSession.Rows[0];
-            _reservationId = Convert.ToInt32(session["Reservation_ID"]);
+            _reservationId = Convert.ToInt64(session["Reservation_ID"]);
             _guestMobilePhone = session["Customer_MobilePhone"].ToString();
-            _accommodationId = Convert.ToInt32(session["Accommodation_ID"]);
+            _accommodationId = Convert.ToByte(session["Accommodation_ID"]);
 
             return true;
         }
